refactor: :recycle: use long for overflow-prone functions

+++ src/core/src/main/scala/com/daromi/cascade/core/math/Primes.scala
@@ -2,17 +2,17 @@ package com.daromi.cascade.core.math
 
 import scala.annotation.tailrec
 
-def isPrime1(n: Int): Boolean =
-  require(n > 1, "n must be greater than 1")
+def isPrime1(n: Long): Boolean =
+  require(n > 1L, "n must be greater than 1")
 
-  if n == 2 || n == 3 then true
+  if n == 2L || n == 3L then true
   else
-    val limit = math.sqrt(n).toInt
+    val limit = math.sqrt(n).toLong
 
     @tailrec
-    def helper(divisor: Int): Boolean =
+    def helper(divisor: Long): Boolean =
       if divisor > limit then true
-      else if n % divisor == 0 then false
-      else helper(divisor + 1)
+      else if n % divisor == 0L then false
+      else helper(divisor + 1L)
 
-    helper(2)
+    helper(2L)
